chore: remove js remnants
