+++ src/components/analyze-forms/ngs2codfreq-form/index.test.ts
@@ -1,20 +1,17 @@
-/*
 import {reformCodFreqs} from './index';
 
-// This test is commented out due to environment constraints when importing
-// ngs2codfreq dependencies which are not yet migrated to TypeScript.
 it('reforms codfreq structure', () => {
   const input = [{
     name: 'sample.codfreq',
-    allReads: [{allCodonReads: [{codon: 'AAA', reads: 1}], gene: 'g', position: 1}]
+    allReads: [
+      {allCodonReads: [{codon: 'AAA', reads: 1}], gene: 'g', position: 1}
+    ]
   }];
   const result = reformCodFreqs(input as any, (g: string, p: number) => [g, p]);
   expect(result[0].name).toBe('sample');
-  expect(result[0].allReads[0].allCodonReads[0]).toEqual({codon: 'AAA', reads: 1});
-});
-*/
-
-it('placeholder', () => {
-  expect(true).toBe(true);
+  expect(result[0].allReads[0].allCodonReads[0]).toEqual({
+    codon: 'AAA',
+    reads: 1
+  });
 });
