+++ src/components/ngs2codfreq/options-form/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+/**
+ * Placeholder options form until fully migrated to TypeScript.
+ *
+ * @param _props - Unused configuration.
+ * @returns Empty fragment.
+ */
+export default function OptionsForm(_props: any): JSX.Element {
+  return <></>;
+}
