+++ src/components/ngs2codfreq/preview-files/drop-placeholder.tsx
@@ -1,58 +1,57 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
 import style from '../style.module.scss';
 
-
-DropPlaceholder.propTypes = {
-  allowFiles: PropTypes.array,
-  blockFiles: PropTypes.array,
-  onMove: PropTypes.func,
-  curDragFile: PropTypes.object,
-  className: PropTypes.string
-};
-
+/**
+ * Placeholder element used while dragging FASTQ files for reordering.
+ *
+ * @param props - Configuration controlling drop behaviour.
+ * @returns List item acting as a drop target.
+ */
 export default function DropPlaceholder({
   allowFiles,
   blockFiles,
   onMove,
   curDragFile,
   className
-}) {
-  // console.log(curDragFile);
-  const allowDrop = (
+}: {
+  allowFiles?: any[];
+  blockFiles?: any[];
+  onMove: (payload: any) => void;
+  curDragFile?: any;
+  className?: string;
+}): JSX.Element {
+  const allowDrop =
     curDragFile &&
     (allowFiles ? allowFiles.includes(curDragFile) : true) &&
-    (blockFiles ? !blockFiles.includes(curDragFile) : true)
-  );
+    (blockFiles ? !blockFiles.includes(curDragFile) : true);
+
   const handleDragOver = React.useCallback(
-    event => {
+    (event: React.DragEvent<HTMLLIElement>) => {
       if (allowDrop) {
         event.preventDefault();
         event.dataTransfer.dropEffect = 'move';
-        event.currentTarget.dataset.dropping = true;
+        (event.currentTarget.dataset as any).dropping = 'true';
       }
     },
     [allowDrop]
   );
 
-  const handleDragLeave = React.useCallback(
-    event => {
-      event.currentTarget.dataset.dropping = false;
-    },
-    []
-  );
+  const handleDragLeave = React.useCallback((event: React.DragEvent<HTMLLIElement>) => {
+    delete (event.currentTarget.dataset as any).dropping;
+  }, []);
 
   const handleDrop = React.useCallback(
-    event => {
+    (event: React.DragEvent<HTMLLIElement>) => {
       event.dataTransfer.dropEffect = 'move';
-      event.currentTarget.dataset.dropping = false;
+      delete (event.currentTarget.dataset as any).dropping;
       const payload = JSON.parse(event.dataTransfer.getData('text'));
       onMove(payload);
     },
     [onMove]
   );
+
   return (
     <li
      className={classNames(
@@ -62,6 +61,7 @@ export default function DropPlaceholder({
      data-await-dropping={allowDrop}
      onDragOver={handleDragOver}
      onDragLeave={handleDragLeave}
-     onDrop={handleDrop} />
+     onDrop={handleDrop}
+    />
   );
 }
